Fixes to tibiaCoins, also fixes from TFS
Fixed tibiaCoins to use int32_t to match the previously existing
IOAccount functions to prevent mismatch later (not that anyone should
ever achieve this number of coins, but just incase). Also added 2 fixes
from TFS to improve performance/reliability and standardize db queries
for player loading.

+++ src/player.cpp
@@ -4118,7 +4118,7 @@ void Player::setPremiumDays(int32_t v)
 	sendBasicData();
 }
 
-void Player::setTibiaCoins(uint64_t v)
+void Player::setTibiaCoins(int32_t v)
 {
 	tibiaCoins = v;
 }
